requirements.txt
cython==0.19.1
setproctitle
-e git+git://github.com/surfly/gevent.git@1.0rc3#egg=gevent
redis==2.7.6
python-dateutil
pytz
